requirements.txt
babelfish>=0.6.0
blessed>=1.19.0
colorama>=0.4.0
docutils>=0.16
guessit>=3.0.0
lockfile==0.12.2
netifaces==0.10.5
pyaml>=21.0.0
python-daemon>=2.3.0
python-dateutil>=2.8.0
PyYAML>=5.3.0
qtfaststart==1.8
rebulk>=3.0.0
six>=1.15.0
libtorrent>=2.0.11